modify for android extension

+++ build/android/gyp/compile_resources.py
@@ -358,6 +358,8 @@ def _MoveImagesToNonMdpiFolders(res_root, path_info):
         continue
       src_file = os.path.join(src_dir, src_file_name)
       dst_file = os.path.join(dst_dir, src_file_name)
+      if os.path.lexists(dst_file):
+          continue
       assert not os.path.lexists(dst_file)
       shutil.move(src_file, dst_file)
       path_info.RegisterRename(
